requirements.txt
asgiref==3.2.10
astroid==2.4.2
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
